Swaped func-s and changed argument types in ctor

+++ Core/Src/main_target.h
@@ -16,7 +16,7 @@ struct LED
     void (* curr_step)(struct LED*);
 };
 
-void ctor_LED(struct LED* led, int detailyty, int pin);
+void ctor_LED(struct LED* led, uint16_t detailyty, uint16_t pin);
 
 void my_delay(int mc_s);
